change field weightInGrammes from Integer to Double

+++ src/main/java/com/orderandwarehouse/app/model/Component.java
@@ -30,7 +30,7 @@ public class Component {
     @Convert(converter = DimensionsConverter.class)
     @Column(length = 11)
     private Dimensions packageDimensions;
-    private Integer weightInGrammes;
+    private Double weightInGrammes;
     @Column(length = 40)
     private String manufacturerId;
     @Column(length = 40)
